requirements.txt
Flask
Flask-PyMongo
Flask-CORS
python-dotenv
pymongo
dnspython
